requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.1
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
pypdf>=4.2.0
langchain_huggingface
chromadb
arxiv
wikipedia
duckduckgo-search
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub #for API calls